pyproject.toml
[project]
name = "woodworking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hexlet-immutable-fs-trees>=0.1.8",
]
